config for realase
