+++ ANDROID-MAC-HACK-KI/app/src/androidTest/java/com/amigotrip/amigo/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.zimincom.amigo
+package com.amigotrip.amigo
 
 import android.support.test.InstrumentationRegistry
 import android.support.test.runner.AndroidJUnit4
