+++ ANDROID-MAC-HACK-KI/app/src/main/java/com/amigotrip/amigo/activities/DoneActivity.kt
@@ -1,8 +1,8 @@
-package com.zimincom.amigo.activities
+package com.amigotrip.amigo.activities
 
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
-import com.zimincom.amigo.R
+import com.amigotrip.amigo.R
 import kotlinx.android.synthetic.main.activity_done.*
 
 class DoneActivity : AppCompatActivity() {
