+++ ANDROID-MAC-HACK-KI/app/src/main/java/com/amigotrip/amigo/activities/IntroActivity.kt
@@ -1,13 +1,13 @@
-package com.zimincom.amigo.activities
+package com.amigotrip.amigo.activities
 
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
-import com.zimincom.amigo.R
+import com.amigotrip.amigo.R
 
-class MessagesActivity : AppCompatActivity() {
+class IntroActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_messages)
+        setContentView(R.layout.activity_intro)
     }
 }
