+++ ANDROID-MAC-HACK-KI/app/src/main/java/com/amigotrip/amigo/adpaters/IntroPagerAdapter.kt
@@ -1,10 +1,10 @@
-package com.zimincom.amigo.adpaters
+package com.amigotrip.amigo.adpaters
 
 import android.support.v4.app.Fragment
 import android.support.v4.app.FragmentManager
 import android.support.v4.app.FragmentStatePagerAdapter
-import com.zimincom.amigo.fragments.IntroPage1Fragment
-import com.zimincom.amigo.fragments.IntroPage2Fragment
+import com.amigotrip.amigo.fragments.IntroPage1Fragment
+import com.amigotrip.amigo.fragments.IntroPage2Fragment
 
 /**
  * Created by Zimincom on 2017. 10. 22..
