+++ ANDROID-MAC-HACK-KI/app/src/main/java/com/amigotrip/amigo/fragments/IntroPage1Fragment.kt
@@ -1,4 +1,4 @@
-package com.zimincom.amigo.fragments
+package com.amigotrip.amigo.fragments
 
 import android.net.Uri
 import android.os.Bundle
@@ -6,22 +6,29 @@ import android.support.v4.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.zimincom.amigo.R
+import com.amigotrip.amigo.R
 import kotlinx.android.synthetic.main.fragment_intro.*
 
 /**
  * Created by Zimincom on 2017. 10. 22..
  */
-class IntroPage2Fragment : Fragment() {
+class IntroPage1Fragment : Fragment() {
+
     override fun onCreateView(inflater: LayoutInflater?, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         return inflater?.inflate(R.layout.fragment_intro, container, false)
     }
 
     override fun onViewCreated(view: View?, savedInstanceState: Bundle?) {
 
-        val uri = Uri.parse("android.resource://" + context.packageName + "/" + R.raw.video_intro_2)
+        val uri = Uri.parse("android.resource://" + context.packageName + "/" + R.raw.intro)
 
         video_intro.setVideoURI(uri)
+        video_intro.isSoundEffectsEnabled = false
         video_intro.start()
+    }
+
+    override fun setUserVisibleHint(isVisibleToUser: Boolean) {
+        super.setUserVisibleHint(isVisibleToUser)
+
     }
 }
