+++ ANDROID-MAC-HACK-KI/app/src/main/res/layout/activity_done.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.zimincom.amigo.activities.DoneActivity">
+    tools:context="com.amigotrip.amigo.activities.DoneActivity">
 
     <Button
         android:id="@+id/button2"
