+++ ANDROID-MAC-HACK-KI/app/src/main/res/layout/activity_parties.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.zimincom.amigo.activities.PartiesActivity">
+    tools:context="com.amigotrip.amigo.activities.PartiesActivity">
 
     <TextView
         android:id="@+id/tv_result"
