+++ ANDROID-MAC-HACK-KI/app/src/main/res/layout/activity_profile.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.zimincom.amigo.activities.ProfileActivity">
+    tools:context="com.amigotrip.amigo.activities.ProfileActivity">
 
 </android.support.constraint.ConstraintLayout>
