+++ ANDROID-MAC-HACK-KI/app/src/test/java/com/amigotrip/amigo/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.zimincom.amigo
+package com.amigotrip.amigo
 
 import org.junit.Test
